Move getCommentValue from plugin comments into shared AST helpers
